Adapting code to make Store requests
This code is base on the https://github.com/alrevuelta/waku-publisher
repo.

With this commit, we have a binary that creates a go-waku node and
connects to two Store nodes (SQLite and Postgres.) and then, it
performs periodic Store requests to both nodes. This is suitable to
compare Store performance between SQLite and Postgres.